Sets up working demo of a broken up api that gives application.py access to housinginsights and demos setting up the structure of breaking endpoints into multiple files inside the /api folder using Blueprints.
