+++ python/housinginsights/tools/test_util.py
@@ -1,0 +1,3 @@
+
+
+api_demo_variable = "Hi I'm inside the Housing Insights package!"
