+++ python/requirements.txt
@@ -15,8 +15,6 @@ jmespath==0.9.2
 MarkupSafe==1.0
 mimerender==0.6.0
 nose==1.3.7
-numpy==1.11.2
-pandas==0.19.0
 psycopg2==2.7.1
 pyasn1==0.2.3
 python-dateutil==2.6.1
@@ -33,5 +31,5 @@ SQLAlchemy==1.1.11
 Werkzeug==0.12.2
 
 #This is only used in local dev, not production (large memory usage and only used for make_draft_json)
-numpy==1.11.2
-pandas==0.19.0
+# numpy==1.11.2
+# pandas==0.19.0
